// frontend/src/app/api/models/medio-catalogacion-entrada.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MedioCatalogacionEntrada {
  autor_credito?: (string | null);
  fuente_url?: (string | null);
  licencia_id?: (number | null);
  pie_de_foto?: (string | null);
  texto_alternativo?: (string | null);
  titulo_interno?: (string | null);
}